+++ src/v3/info.ts
@@ -6,6 +6,7 @@ import type { InfoDocInput, InfoDoc } from '../contracts';
 
 export const profileLink = 'https://iiif.io/api/image/3/level2.json';
 
+const DEFAULT_TILE_SIZE = 512;
 const defaultFormats: Set<string> = new Set(['jpg', 'png']);
 const defaultQualities: Set<string> = new Set(['default']);
 const IIIFExtras = {
@@ -23,19 +24,14 @@ const IIIFExtras = {
   extraQualities: new Set(Qualities.filter((q) => !defaultQualities.has(q)))
 };
 
-export function infoDoc({
-  id,
-  width,
-  height,
-  sizes,
-  max
-}: InfoDocInput): InfoDoc {
+export function infoDoc({ id, geometry, max }: InfoDocInput): InfoDoc {
   const maxAttrs = {
     maxWidth: max?.width,
     maxHeight: max?.height,
     maxArea: max?.area
   };
 
+  const { width, height, sizes } = geometry;
   return {
     '@context': 'http://iiif.io/api/image/3/context.json',
     id,
@@ -47,8 +43,8 @@ export function infoDoc({
     sizes,
     tiles: [
       {
-        width: 512,
-        height: 512,
+        width: geometry.tileWidth || DEFAULT_TILE_SIZE,
+        height: geometry.tileHeight || DEFAULT_TILE_SIZE,
         scaleFactors: sizes.map((_v: Dimensions, i: number) => 2 ** i)
       }
     ],